+++ R/gdal.R
@@ -27,7 +27,8 @@ gdal <- function(warn=NA, drivers=FALSE) {
 		x <- .gdaldrivers()
 		x <- do.call(cbind, x)
 		x[,2] = c("vector", "raster")[as.integer(x[,2])+1]
-		x[,3] = c("read", "read/copy-write", "read/write")[as.integer(x[,3])+1]
+		#x[,3] = c("read", "read/copy-write", "read/write")[as.integer(x[,3])+1]
+		x[,3] = c("read", "read/write", "read/write")[as.integer(x[,3])+1]
 		colnames(x) <- c("name", "type", "can", "vsi", "long.name")
 		x <- data.frame(x)
 		x[,4] <- x[,4] == 1